f lvl api test prog

+++ api/v1/views/comment.py
@@ -2,6 +2,7 @@
 
 from api.v1.views import app_views
 from models.comment import Comment
+from models.post import Post
 from models import db, app
 from api.v1.auth.jwt_auth import jwt_required
 from flask_jwt_extended import get_current_user
@@ -22,18 +23,20 @@ def create_comment(id, post_id):
     """
     current_user = get_current_user()
     if not request.json:
-        abort(404)
+        abort(400)
     allowed_fields = ['message']
     for field in allowed_fields:
         if field not in request.json:
-            abort(404)
+            abort(400)
     data = request.get_json()
-    data['user_id'] = id
+    with app.app_context():
+        post = db.one_or_404(db.select(Post).filter_by(id=post_id))
+    data['user_id'] = post.user_id
     data['post_id'] = post_id
     data['commenter_id'] = current_user.id
     comment = Comment(**data)
     if comment is None:
-        abort(404)
+        abort(400)
     with app.app_context():
         db.session.add(comment)
         db.session.commit()
@@ -43,9 +46,9 @@ def create_comment(id, post_id):
                                   'post_id': comment.post_id}), 201)
 
 
-@app_views.route('/post/update/<comment_id>', methods=['PUT'], strict_slashes=False)
+@app_views.route('/comment/update/<comment_id>', methods=['PUT'], strict_slashes=False)
 @jwt_required()
-def update_comment(id, comment_id):
+def update_comment(comment_id):
     """
     updates a comment with a new message and returns the updated comment
     details in a JSON response.
@@ -54,7 +57,7 @@ def update_comment(id, comment_id):
     # if current_user.id != id:
     #     abort(401)
     if not request.json:
-        abort(404)
+        abort(400)
     allowed_fields = ['message']
     for field in request.json:
         if field not in allowed_fields:
@@ -62,6 +65,8 @@ def update_comment(id, comment_id):
     data = request.get_json()
     with app.app_context():
         comment = db.get_or_404(Comment, comment_id)
+        if comment.commenter_id != current_user.id:
+            abort(401)
         if 'message' in data:
             comment.message = data['message']
         db.session.commit()
@@ -88,6 +93,8 @@ def delete_comment(id, comment_id):
         abort(404)
     with app.app_context():
         comment = db.get_or_404(Comment, comment_id)
+        if comment.commenter_id != current_user.id:
+            abort(401)
         db.session.delete(comment)
         db.session.commit()
     return make_response(jsonify({'deleted': True}), 200)
