+++ api/v1/views/post.py
@@ -4,7 +4,7 @@ from api.v1.views import app_views
 from models.post import Post
 from models import db, app
 from api.v1.auth.jwt_auth import jwt_required
-from flask_jwt_extended import get_current_user
+from flask_jwt_extended import get_current_user, get_jwt_identity
 from flask import (
                     request,
                     make_response,
@@ -73,6 +73,8 @@ def update_post(id, post_id):
     data = request.get_json()
     with app.app_context():
         post = db.get_or_404(Post, post_id)
+        if current_user.id != post.user_id:
+            abort(401)
         if 'title' in data:
             post.title = data['title']
         if 'body' in data:
@@ -112,6 +114,8 @@ def delete_post(id, post_id):
         abort(404)
     with app.app_context():
         post = db.get_or_404(Post, post_id)
+        if current_user.id != post.user_id:
+            abort(401)
         db.session.delete(post)
         db.session.commit()
     return make_response(jsonify({'deleted': True}), 200)
@@ -124,6 +128,8 @@ def get_post(id, post_id):
         abort(401)
     with app.app_context():
         post = db.one_or_404(Post, post_id)
+        if user.id != post.user_id:
+            abort(401)
         all_comments = post.comments
         comments = [
                     {
@@ -137,3 +143,31 @@ def get_post(id, post_id):
                                   'author': post.author,
                                   'comments': comments,
                                   'publication_date': post.publication_date}), 200)
+
+
+@app_views.route('/post/<id>', methods=['GET'], strict_slashes=False)
+@jwt_required()
+def get_all_posts(id):
+    user = get_current_user()
+    if user.id != id:
+        abort(401)
+    with app.app_context():
+        all_posts = [p for p in user.posts if p.user_id == user.id]
+        i = 0
+        for post in all_posts:
+            comments = [{
+                            'message': c.message, 
+                            'comment_date': c.comment_date,
+                            'commenter_id': c.commenter_id, 
+                            'post_id': c.post_id} for c in post.comments]
+            post_dict = {'title': post.title,
+                                  'body': post.body,
+                                  'author': post.author,
+                                  'comments': comments,
+                                  'publication_date': post.publication_date}
+            
+            all_posts[i] = post_dict
+            i += 1
+            # print(all_posts[i - 1])
+
+    return make_response(jsonify(all_posts), 200)
